[WIP] Move Mesh class from tests to code base
Unittest tests/test_mesh.py working
